Increase timeline chart buffer.
Reduce action_t::init(); calls in action ctor's by constructing the
actions earlier.

Fix warnings in rng about ambigious sqrt call.

Move player callbacks and their register functions into one consolidated
container.

DPS Error timeline: Only display graph from n>=50 onwards.

git-svn-id: https://simulationcraft.googlecode.com/svn/branches/mop@11619 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_rng.cpp
@@ -133,9 +133,9 @@ double rng_t::exgauss( double mean, double stddev, double nu )
 std::string rng_t::report( double confidence_estimator ) const
 {
   double nan = std::numeric_limits<double>::quiet_NaN();
-  double gauss_confidence = 1.0 / sqrt( num_gauss ) * confidence_estimator;
-  double range_confidence = 1.0 / sqrt( num_range ) * confidence_estimator;
-  double roll_confidence = 1.0 / sqrt( num_roll ) * confidence_estimator;
+  double gauss_confidence = 1.0 / sqrt( static_cast<double>( num_gauss ) ) * confidence_estimator;
+  double range_confidence = 1.0 / sqrt( static_cast<double>( num_range ) ) * confidence_estimator;
+  double roll_confidence = 1.0 / sqrt( static_cast<double>( num_roll ) ) * confidence_estimator;
 
   int precision = 6;
